tools/e2e-test/requirements.txt
requests>=2.31.0
pyyaml>=6.0
pillow>=10.0.0
google-generativeai>=0.3.0
click>=8.1.0
